requirements.txt
fastapi[standard]==0.115.14
sqlmodel==0.0.24
python-dotenv==1.1.1
psycopg2-binary==2.9.10
